Add more namespaces
