+++ decidim-ai/lib/decidim/ai/spam_detection/strategy/base.rb
@@ -3,23 +3,25 @@
 module Decidim
   module Ai
     module SpamDetection
-      class Base
-        attr_reader :name
+      module Strategy
+        class Base
+          attr_reader :name
 
-        def initialize(options = {})
-          @name = options.delete(:name)
-          @options = options
+          def initialize(options = {})
+            @name = options.delete(:name)
+            @options = options
+          end
+
+          def classify(_content); end
+
+          def train(_classification, _content); end
+
+          def untrain(_classification, _content); end
+
+          def log; end
+
+          def score = 0.0
         end
-
-        def classify(_content); end
-
-        def train(_classification, _content); end
-
-        def untrain(_classification, _content); end
-
-        def log; end
-
-        def score = 0.0
       end
     end
   end
