+++ decidim-ai/lib/tasks/decidim_ai.rake
@@ -10,13 +10,13 @@ namespace :decidim do
     desc "Load plugin shipped datasets"
     task load_plugin_dataset: :environment do
       Dir.glob("#{plugin_path}/data/*.csv").each do |file|
-        Decidim::Ai::LoadDataset.call(file)
+        Decidim::Ai::SpamDetection::Importer::File.call(file)
       end
     end
 
     desc "Load application datasets"
     task :load_application_dataset, [:file] => :environment do |_, args|
-      Decidim::Ai::LoadDataset.call(args[:file])
+      Decidim::Ai::SpamDetection::Importer::File.call(args[:file])
     end
 
     private
