+++ decidim-ai/spec/event_handlers/user/user_changes_profile_data_spec.rb
@@ -54,7 +54,7 @@ describe "User changes own data", type: :system do
   before do
     Decidim::Ai.spam_detection_registry.clear
     Decidim::Ai.spam_detection_registry.register_analyzer(name: :bayes,
-                                                          strategy: Decidim::Ai::SpamContent::BayesStrategy,
+                                                          strategy: Decidim::Ai::SpamDetection::Strategy::Bayes,
                                                           options: { adapter: :memory, params: {} })
 
     Decidim::Ai.spam_detection_instance.train :ham, "I am a passionate Decidim Maintainer. It is nice to be here."
