+++ decidim-ai/spec/lib/decidim/ai/spam_detection/importer/file_spec.rb
@@ -2,12 +2,12 @@
 
 require "spec_helper"
 
-describe Decidim::Ai::LoadDataset do
+describe Decidim::Ai::SpamDetection::Importer::File do
   let!(:path) { Gem.loaded_specs["decidim-ai"].full_gem_path }
   let!(:file) { "spec/support/test.csv" }
 
   it "successfully loads the dataset" do
-    instance = Decidim::Ai::SpamDetectionService.new
+    instance = Decidim::Ai::SpamDetection::Service.new
     allow(Decidim::Ai).to receive(:spam_detection_instance).and_return(instance)
     expect(instance).to receive(:train).exactly(4).times
