+++ decidim-ai/spec/lib/decidim/ai/spam_detection/strategy/bayes_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe Decidim::Ai::SpamDetection::Bayes do
+describe Decidim::Ai::SpamDetection::Strategy::Bayes do
   subject { described_class.new({}) }
 
   let(:backend) { ClassifierReborn::Bayes.new :spam, :ham }
